Window stack: Moved the insertion of windows into another event method
Some clients, such as OpenGL ones (for example, `alacritty`), create
multiple windows that do not actually request to be mapped. Thus, having
the insertion of windows into the stack in the `CreateNotify` method,
could potentially create a lot of unnecessary entries that cannot be
actually seen, interacted with, and focused by cycling the stack.

Putting the creation of `Client` objects in the `MapRequest` mitigates
this, in turn making the window stack MUCH cleaner and actually useful.

I am not sure about how this will, or IF it will, impact functionality
in the future.