out_of_bounds condition found.
We try to focus a nonexistent ID '1' window. This happens because
XGetInputFocus() apparently stores '1' in the variable passed by
reference when there's no active focused window. This can be caused by
closing the focused window, and because we don't refocus to another
active window, if you try to KillClient when you do NOT refocus, the
program will grind to a halt.

We've found a solution, that is, refocusing to another window when the
currently focused one gets killed. This would be easier if I'd already
implemented the tiling funcionality, because I'd have a stack to rely
on. This probably won't get fixed until the tiling gets implemented.